separated creation from deletion

+++ Windows/deleteStructure.ps1
@@ -1,5 +1,9 @@
 $list=Get-ADOrganizationalUnit -Filter 'Name -like "*"' | Format-Table Name -A
+$toRemove= "Sites" , "T0" , "T1" , "T2"
 foreach ($item in $list) {
-    Set-ADOrganizationalUnit -ProtectedFromAccidentalDeletion $false -Identity $item
-    Remove-ADOrganizationalUnit -Identity $item -Confirm:$false
+    if ($toRemove -contains $item) {
+        Write-Host "Removing $item"
+        Set-ADOrganizationalUnit -ProtectedFromAccidentalDeletion $false -Identity $item
+        Remove-ADOrganizationalUnit -Identity $item -Confirm:$false
+    }
 }
